fix: give correct identifier to app
This fixes a bug where more than two components with the same name could
not be held apart by the backend. It is looking for `defaultId:count`
while counting references but then names the reference `defaultId_count`
this resulted in a maximum of two components with the same name being
recognized (that is clickable) in the devtools.