fix(chrome): stop extension unloading in Firefox, fix #1974
Firefox starting with version 106 introduced a change which
apparently broke background pages. For the Vuejs devtools, this
surfaced as extension unloading when the browser reached an
"idle" state. User having the extension open noticed it as
a unloading/reloading of the devtools tab.

By setting background.persisted to true in the manifest.json,
the unloading stopped.

Note: This is a temporary solution, until Manifest v3 is
properly supported.